[SCRUM-54] feat: Refactor - Storage 변경에 따른 Clothe 관련 코드 수정
