+++ src/main/java/com/sprint/otboo/clothing/service/impl/ClothesServiceImpl.java
@@ -12,8 +12,8 @@ import com.sprint.otboo.clothing.repository.ClothesAttributeDefRepository;
 import com.sprint.otboo.clothing.repository.ClothesAttributeRepository;
 import com.sprint.otboo.clothing.repository.ClothesRepository;
 import com.sprint.otboo.clothing.service.ClothesService;
-import com.sprint.otboo.clothing.storage.FileStorageService;
 import com.sprint.otboo.common.dto.CursorPageResponse;
+import com.sprint.otboo.common.storage.FileStorageService;
 import com.sprint.otboo.user.repository.UserRepository;
 import java.time.Instant;
 import java.util.List;
