+++ src/main/java/com/sprint/otboo/clothing/storage/FileStorageService.java
@@ -1,10 +1,10 @@
-package com.sprint.otboo.clothing.storage;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public interface FileStorageService {
-
-    String upload(MultipartFile file);
-
-    void delete(String url);
-}
+//package com.sprint.otboo.clothing.storage;
+//
+//import org.springframework.web.multipart.MultipartFile;
+//
+//public interface FileStorageService {
+//
+//    String upload(MultipartFile file);
+//
+//    void delete(String url);
+//}
