+++ src/test/java/com/sprint/otboo/clothing/service/ClothesServiceTest.java
@@ -23,8 +23,8 @@ import com.sprint.otboo.clothing.repository.ClothesAttributeDefRepository;
 import com.sprint.otboo.clothing.repository.ClothesAttributeRepository;
 import com.sprint.otboo.clothing.repository.ClothesRepository;
 import com.sprint.otboo.clothing.service.impl.ClothesServiceImpl;
-import com.sprint.otboo.clothing.storage.FileStorageService;
 import com.sprint.otboo.common.dto.CursorPageResponse;
+import com.sprint.otboo.common.storage.FileStorageService;
 import com.sprint.otboo.feed.entity.FeedClothes;
 import com.sprint.otboo.user.entity.User;
 import com.sprint.otboo.user.repository.UserRepository;
